Issue #132: Moved setup.php definitions to setuplib.php so setup.php can be required more than once.
